refactor : vo 패키지 constant 로 리펙터링
