+++ src/main/java/lotto/constant/Grade.java
@@ -1,4 +1,4 @@
-package lotto.vo;
+package lotto.constant;
 
 public enum Grade {
