+++ src/main/java/lotto/model/Lotto.java
@@ -1,8 +1,8 @@
 package lotto.model;
 
-import static lotto.vo.CommonUnits.MAX_LOTTO_NUMBER;
-import static lotto.vo.CommonUnits.MIN_LOTTO_NUMBER;
-import static lotto.vo.CommonUnits.NUMBERS_OF_LOTTO;
+import static lotto.constant.CommonUnits.MAX_LOTTO_NUMBER;
+import static lotto.constant.CommonUnits.MIN_LOTTO_NUMBER;
+import static lotto.constant.CommonUnits.NUMBERS_OF_LOTTO;
 
 import java.util.List;
 import java.util.stream.Collectors;
